Fix altas facing sprite particle fraction selection, fix more screen layering issues, adjust size of aoe preview to match the actual size more accurately, add tree beacon, add tree beacon patreon effecttype, generalize faked state handling

+++ src/main/java/hellfirepvp/astralsorcery/client/effect/function/VFXColorFunction.java
@@ -52,9 +52,9 @@ public interface VFXColorFunction<T extends EntityVisualFX> {
             @Nonnull
             @Override
             public Color getColor(@Nonnull T fx, float pTicks) {
-                float h = (float) RenderingVectorUtils.interpolate(hsb1[0], hsb2[0], degree);
-                float s = (float) RenderingVectorUtils.interpolate(hsb1[1], hsb2[1], degree);
-                float b = (float) RenderingVectorUtils.interpolate(hsb1[2], hsb2[2], degree);
+                float h = RenderingVectorUtils.interpolate(hsb1[0], hsb2[0], degree);
+                float s = RenderingVectorUtils.interpolate(hsb1[1], hsb2[1], degree);
+                float b = RenderingVectorUtils.interpolate(hsb1[2], hsb2[2], degree);
                 return Color.getHSBColor(h, s, b);
             }
         };
